pyproject.toml
[tool.poetry]
name = "lab-devices"
version = "0.1.0"
description = "Async API for laboratory hardware devices with serial port interfaces"
authors = []
packages = [{include = "lab_devices", from = "src"}]

[tool.poetry.dependencies]
python = "^3.14"
aioserial = "^1.3"
pyserial = "^3.5"
pydantic = "^2.0"
pydantic-settings = {version = "^2.6", extras = ["yaml"]}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.25"
mypy = "^1.0"
ruff = "^0.9"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["aioserial", "serial", "serial.*", "serial.tools.*"]
ignore_missing_imports = true
follow_untyped_imports = false

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
